mail_service/pyproject.toml
[tool.poetry]
name = "mail_service"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
gino = {extras = ["starlette"], version = "^1.0.1"}
fastapi = "^0.60.2"
uvicorn = "^0.11.8"
gunicorn = "^20.0.4"
alembic = "^1.4.2"
psycopg2 = "^2.8.5"
aiosmtplib = "^1.1.3"
asyncio = "^3.4.3"
python-dotenv = "^0.14.0"
isort = "^5.4.2"
flake8 = "^3.8.3"
black = "^19.10b0"
aio-pika = "^6.8.0"

[tool.poetry.dev-dependencies]

[tool.poetry.plugins."mail_service.modules"]
"mails" = "mail_service.views.mails"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

